Refactoring. Tests.
